requirements.txt
Django==1.11.29
django-cors-middleware==1.3.1
django-extensions==1.7.1
djangorestframework==3.4.4
PyJWT==1.4.2
six==1.10.0
uptrace
opentelemetry-instrumentation-django
opentelemetry-instrumentation-psycopg2
